Added shell script to help with starting the server.
Updated the project to allow use of local language models for transcribing and summarizing. Local transcription is done with faster-whisper, while users can configure information for a local server with OpenAI API-like endpoints.

Editing summaries with a language model now also uses structured output, to ensure consistency and avoid odd or unwanted text.

Updates and fixes to templates.

Added some test audio files and testing observations.

+++ transcriber/be/django_project/transcriber/gpt_transcription.py
@@ -5,30 +5,47 @@ from openai import OpenAI
 from pydantic import BaseModel
 from pydub import AudioSegment
 from pydub.silence import detect_silence
+from faster_whisper import WhisperModel
 
-from django_project.settings import OPENAI_API_KEY
+from django_project.settings import OPENAI_API_KEY, LOCAL_LLM_API_PORT
+from transcriber.models import Transcription
 
 
 # Get logger
 logger = logging.getLogger(__name__)
 
 # Initialize OpenAI client
-client = OpenAI(api_key=OPENAI_API_KEY)
+openai_client = OpenAI(api_key=OPENAI_API_KEY)
+local_client = OpenAI(
+    api_key='',
+    base_url=f'http://127.0.0.1:{LOCAL_LLM_API_PORT}/v1',
+)
 
-PROMPT_TRANSCRIPTION = ''.join([
-    "You are a medical transcriber. ",
-    "You expect a transcript of a doctor-patient conversation. ",
+# OpenAI models
+OPENAI_GPT_4O_MINI_TRANSCRIBE = "gpt-4o-mini-transcribe"
+
+# Whisper model configuration
+WHISPER_LARGE_V3_TURBO = "large-v3-turbo"
+whisper_model = WhisperModel(WHISPER_LARGE_V3_TURBO, device="cpu", compute_type="int8")
+
+PROMPT_SUMMARIZATION = ''.join([
+    "You are a medical assistant for summarizing patient encounters. ",
+    "You expect a transcript of a doctor-patient conversation or related information. ",
     "If the user provides it, please summarize the transcript into a SOAP format note. ",
-    "If it is not a valid transcript, leave the fields empty.",
+    "If it is not a valid transcript, leave the fields empty. ",
+    "If there is insufficient information, do not make up information. ",
+    "Do not make up any details not found in the transcript. ",
     "Use professional medical terminology. ",
     "Use basic HTML styling. Do not wrap in extra strings or quotations. ",
     "Translate to English if transcript is in another language.",
 ])
 
 PROMPT_EDIT = ''.join([
-    "You are a medical text editing assistant. ",
+    "You are a medical assistant for editing encounter documentation.",
     "You will be given a SOAP format medical note and instructions on editing it. ",
-    "Update and return the SOAP format note. Use HTML formatting.",
+    "Update the summary according to instructions.",
+    "Use HTML styling consistent with the input, if formatting is requested. ",
+    "Do not wrap in extra strings or quotations.",
 ])
 
 CHUNK_SIZE_MB = 35      # Actual stored file size around 3 MB when set to 35
@@ -41,8 +58,8 @@ class SOAPNote(BaseModel):
     The SOAP format is as follows:
     - Subjective: The patient's subjective medical history.
     - Objective: The patient's objective medical history.
-    - Assessment: The patient's assessment.
-    - Plan: The patient's plan.
+    - Assessment: The assessment of the patient.
+    - Plan: The plan for the patient.
     """
     subjective: str
     objective: str
@@ -80,10 +97,10 @@ class SOAPNote(BaseModel):
     def to_html_with_headers(self):
         """Output as a single HTML string with headers."""
         return '<br><br>'.join([
-            f'Subjective:<br>{self.subjective}',
-            f'Objective:<br>{self.objective}',
-            f'Assessment:<br>{self.assessment}',
-            f'Plan:<br>{self.plan}',
+            f'<b>Subjective:</b><br>{self.subjective}',
+            f'<b>Objective:</b><br>{self.objective}',
+            f'<b>Assessment:</b><br>{self.assessment}',
+            f'<b>Plan:</b><br>{self.plan}',
         ])
 
 
@@ -172,32 +189,51 @@ def split_mp3_to_chunks(mp3_path, chunk_size_mb):
     return chunks
 
 
-def get_transcription_from_local_file(path):
+def get_transcription_from_local_file(path: str, model_choice: str = WHISPER_LARGE_V3_TURBO):
     """Get transcription from OpenAI API."""
-    chunks = split_mp3_to_chunks(path, CHUNK_SIZE_MB)
-
     transcripts = []
-    for idx, chunk_path in enumerate(chunks):
-        logger.info("Transcribing chunk %s/%s: %s", idx + 1, len(chunks), chunk_path)
+    logger.info('transcribing with model: %s', model_choice)
 
-        # Transcribe chunk
-        with open(chunk_path, "rb") as audio_file:
-            try:
-                transcription = client.audio.transcriptions.create(
-                    model="gpt-4o-mini-transcribe",
-                    file=audio_file
-                )
-                transcripts.append(transcription.text)
-            except Exception as e:
-                logger.error("Error transcribing %s: %s", chunk_path, e)
-                transcripts.append(f"[ERROR in chunk {chunk_path}]")
-        # Remove chunk file after transcription
-        os.remove(chunk_path)
+    if WHISPER_LARGE_V3_TURBO in model_choice:
+        # Local whisper model transcription
+        logger.info('transcribing with whisper large v3 turbo')
+        segments, info = whisper_model.transcribe(path, beam_size=5)
+        logger.info(
+            "Detected language '%s' with probability %s",
+            info.language,
+            info.language_probability
+        )
+        transcripts.append(' '.join(segment.text for segment in segments))
+        logger.info('transcribed segments')
+        logger.info(segments)
+    else:
+        # Split audio into chunks
+        logger.info('transcribe with gpt')
+        chunks = split_mp3_to_chunks(path, CHUNK_SIZE_MB)
+
+        # Transcribe with OpenAI API transcription
+        for idx, chunk_path in enumerate(chunks):
+            logger.info("Transcribing chunk %s/%s: %s", idx + 1, len(chunks), chunk_path)
+
+            # Transcribe chunk
+            with open(chunk_path, "rb") as audio_file:
+                try:
+                    transcription = openai_client.audio.transcriptions.create(
+                        model="gpt-4o-mini-transcribe",
+                        file=audio_file
+                    )
+                    transcripts.append(transcription.text)
+                except Exception as e:
+                    logger.error("Error transcribing %s: %s", chunk_path, e)
+                    transcripts.append(f"[ERROR in chunk {chunk_path}]")
+
+            # Remove chunk file after transcription
+            os.remove(chunk_path)
 
     return "\n".join(transcripts)
 
 
-def get_soap_format_from_transcription(transcript):
+def get_soap_format_from_transcription(transcript: str, model_choice: str):
     """Get SOAP format from transcription."""
     # Skip in case of empty transcript
     if not transcript.strip():
@@ -208,30 +244,64 @@ def get_soap_format_from_transcription(transcript):
             plan='n/a - no transcript text'
         ).to_html()
 
-    # Get response from OpenAI API
-    response = client.chat.completions.parse(
-        model="gpt-4o-mini",
-        response_format=SOAPNote,
-        messages=[
-            {"role": "system", "content": PROMPT_TRANSCRIPTION},
-            {"role": "user", "content": transcript},
-        ]
-    )
+    logger.info('model_choice: %s', model_choice)
+
+    messages = [
+        {"role": "system", "content": PROMPT_SUMMARIZATION},
+        {"role": "user", "content": transcript},
+    ]
+
+    if 'gpt' not in model_choice:
+        # Get response from local OpenAI-like API
+        response = local_client.chat.completions.parse(
+            model=model_choice,
+            response_format=SOAPNote,
+            messages=messages
+        )
+    else:
+        # Get response from OpenAI API
+        response = openai_client.chat.completions.parse(
+            model=model_choice,
+            response_format=SOAPNote,
+            messages=messages
+        )
 
     parsed_output = response.choices[0].message.parsed
 
     return parsed_output.to_html_with_headers()
 
 
-def update_soap_format_with_intruction(transcription_obj, input_text):
+def update_soap_format_with_instruction(
+    transcription_obj: Transcription,
+    input_text: str,
+    model_choice: str
+):
     """Modify SOAP format with instruction."""
-    # Get response from OpenAI API
-    response = client.chat.completions.create(
-        model="gpt-4o-mini",
-        messages=[
-            {"role": "system", "content": PROMPT_EDIT},
-            {"role": "user", "content": transcription_obj.formatted_text + '\n' + input_text},
-        ]
-    )
+    messages = [
+        {"role": "system", "content": PROMPT_EDIT},
+        {"role": "user", "content": '\n'.join([
+            'Text:',
+            transcription_obj.formatted_text,
+            'Instructions:',
+            input_text
+        ])},
+    ]
 
-    return response.choices[0].message.content
+    if 'gpt' not in model_choice:
+        # Get response from local OpenAI-like API
+        response = local_client.chat.completions.parse(
+            model=model_choice,
+            response_format=SOAPNote,
+            messages=messages
+        )
+    else:
+        # Get response from OpenAI API
+        response = openai_client.chat.completions.parse(
+            model=model_choice,
+            response_format=SOAPNote,
+            messages=messages
+        )
+
+    parsed_output = response.choices[0].message.parsed
+
+    return parsed_output.to_html_with_headers()
